Update button component to replace type string with type enum.
